Fix #112: Update SpyingMIDIDriver `SizeOfPacketList` for Apple Silicon
The old code to calculate the MIDIPacketList's size didn't take into account the 4-byte alignment of `MIDIPacket`s on Apple Silicon.
Copied known-good code from `SMPacketListSize()` in the app.

+++ Frameworks/SnoizeMIDISpy/Driver/SpyingMIDIDriver.h
@@ -28,7 +28,7 @@ private:
     void EnableMonitoring(Boolean enable);
 
     CFMutableDataRef PackageMonitoredDataForBroadcast(MIDIEndpointRef destination, const MIDIPacketList *packetList);
-    UInt32 SizeOfPacketList(const MIDIPacketList *packetList);
+    intptr_t SizeOfPacketList(const MIDIPacketList *packetList);
 
     
     MessagePortBroadcaster *mBroadcaster;
